ci: use blacksmith runners for gh actions (#271)
* ci: use blacksmith 16vcpu runners

This change switches over the `basic` workflow
to Blacksmith's 16vcpu runners. Additionally, this
change switches the rust cache action to use the
Blacksmith drop-in replacement. The Blacksmith version
of the cache is colocated with Blacksmith runners and is
substantially faster.

The result is that this workflow is overall 2x faster and
half the cost.

* ci: move coverage to blacksmith 16vcpu runners

* ci: remove `workflow_dispatch` from the triggers

* ci: bump the `Test Suite` job to use 32vCPUs

* ci: update `setup-protoc` step to avoid getting rate limited

This commit adds an auth token to the `setup-protoc` step to avoid getting rate limited as per: https://github.com/arduino/setup-protoc?tab=readme-ov-file#usage.

* ci: update `Test Suite` job to authenticate for `setup-protoc`

This commit adds an auth token to the `setup-protoc` step to avoid getting rate limited as per: https://github.com/arduino/setup-protoc?tab=readme-ov-file#usage.

* ci: move `Test Suite` job back to 16vCPUs

* ci: move `Test Suite` and compilation steps to 32vCPU VMs

---------

Co-authored-by: Aditya Maru <adityamaru@gmail.com>
Co-authored-by: Aayush Shah <aayush.shah15@gmail.com>